+++ parley/src/layout/mod.rs
@@ -3,8 +3,6 @@
 
 //! Layout types.
 
-#[cfg(feature = "accesskit")]
-mod accessibility;
 mod alignment;
 mod cluster;
 mod line;
@@ -22,8 +20,6 @@ pub use parley_engine::{FontMetrics, Glyph};
 
 pub(crate) mod data;
 
-#[cfg(feature = "accesskit")]
-pub use accessibility::LayoutAccessibility;
 pub use alignment::{Alignment, AlignmentOptions};
 pub use cluster::{Affinity, Cluster, ClusterPath, ClusterSide};
 pub use data::BreakReason;
@@ -62,9 +58,11 @@ pub struct Style<B: Brush> {
     pub(crate) overflow_wrap: OverflowWrap,
     /// Per-cluster text-wrap-mode setting
     pub(crate) text_wrap_mode: TextWrapMode,
-    #[cfg(feature = "accesskit")]
-    /// Locale if any, so we can set the corresponding AccessKit property
-    pub(crate) locale: Option<fontique::Language>,
+    /// The locale of the text, if any.
+    ///
+    /// This is useful for consumers exposing the text to assistive technologies,
+    /// which use it to select an appropriate pronunciation.
+    pub locale: Option<fontique::Language>,
 }
 
 /// Underline or strikethrough decoration.